Fix for latest rust in f25: undefined reference to `pthread_mutex_trylock'
Fix Description: The latest rustc / cargo does not work, as it no longer
links a set of libraries during build. This causes a large number of
undefined reference messages.

Add a LDFLAGS argument that provides the correct link arguments for linux.
Additionally, add a clean-local, allowing the make clean command to work.